Rename lang module to locale
Locale will make more sense if we add in items like timezone.
