+++ mycroft/configuration/__init__.py
@@ -13,5 +13,5 @@
 # limitations under the License.
 #
 from .config import Configuration, LocalConf, RemoteConf
-from .lang import set_default_lf_lang
+from .locale import set_default_lf_lang
 from .locations import SYSTEM_CONFIG, USER_CONFIG
