+++ mycroft/configuration/locale.py
